+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 2024-04-16
 
 - 实现移动端界面语言切换功能 | Implement mobile interface language switching feature
+- 实现移动端界面主题切换功能 | Implement mobile interface theme switching feature
 
 ## 2024-04-15